Split golfer rankings colors to own file
